{{cookiecutter.project_slug}}/requirements.txt
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
boto3==1.35.45
certifi==2024.8.30
click==8.1.7
dependency-injector==4.41.0
Deprecated==1.2.15
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
flake8==7.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
opensearch-py==2.7.1
packaging==24.2
pre-commit==4.0.1
psutil==6.1.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
Pygments==2.18.0
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv==1.0.1
python-keycloak==3.12.0
python-multipart==0.0.17
PyYAML==6.0.2
redis==5.0.7
requests-aws4auth==1.3.1
rich==13.9.4
ruff==0.5.5
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.37.2
testcontainers==4.9.0
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
wrapt==1.17.0
